api implementation refactorized
