+++ src/navigation/AppNavigator/style.js
@@ -1,0 +1,16 @@
+import {StyleSheet} from 'react-native';
+
+export const styles = StyleSheet.create({
+  navigateButton: {
+    margin: 5,
+    backgroundColor: '#bd94ff',
+    borderRadius: 20,
+    padding: 5,
+  },
+  navigateTxt: {
+    padding: 3,
+    fontSize: 15,
+    fontWeight: 'bold',
+    color: '#1f1633',
+  },
+});
